services/infoscreen/requirements.txt
Flask==3.0.0
gpiozero==2.0
matplotlib==3.8.2
Pillow==10.1.0
python-dateutil==2.8.2
requests==2.31.0
RPi.GPIO==0.7.1
spidev==3.6
scipy==1.11.4